requirements.txt
web3
google-genai
python-dotenv
requests
pypdf